+++ src/images/index.js
@@ -2,6 +2,14 @@ import splashGenericErrorSvg from './splash-generic-error.svg';
 import splashSelectFileSvg from './splash-select-file.svg';
 import TarantoolLogoCompact from './tarantool-logo-compact.svg';
 import TarantoolLogoFull from './tarantool-logo-full.svg';
+import TarantoolLogoUser from './tarantool-logo-user.svg';
 import windowErrorSvg from './window-error.svg';
 
-export { splashGenericErrorSvg, splashSelectFileSvg, TarantoolLogoCompact, TarantoolLogoFull, windowErrorSvg };
+export {
+  splashGenericErrorSvg,
+  splashSelectFileSvg,
+  TarantoolLogoCompact,
+  TarantoolLogoFull,
+  TarantoolLogoUser,
+  windowErrorSvg,
+};